refactor(common)!: move thread related functions to their own header
