+++ src/bsa/pack.cpp
@@ -10,6 +10,7 @@
 
 #include <btu/common/algorithms.hpp>
 #include <btu/common/functional.hpp>
+#include <btu/common/threading.hpp>
 #include <flux.hpp>
 
 #include <functional>
