+++ tests/CMakeLists.txt
@@ -11,6 +11,7 @@ set(SOURCE_FILES
     "${SOURCE_DIR}/common/json.cpp"
     "${SOURCE_DIR}/common/metaprogramming.cpp"
     "${SOURCE_DIR}/common/string.cpp"
+    "${SOURCE_DIR}/common/threading.cpp"
     "${SOURCE_DIR}/bsa/archive.cpp"
     "${SOURCE_DIR}/bsa/pack.cpp"
     "${SOURCE_DIR}/bsa/plugin.cpp"
